quantified the successfulness of attack by the percentage of wrong load estimate compared with no attack
Suppose the attacker targets at misleading the operator into thinking the load increases a lot (by 20% the true value)
Then total_Pload_diff = 0.2 means, the operator under attack makes a higher load estimate than the estimate without attack, by 20% of the true load.
This means the attack is very successful.
----------------------------
also if the attack aims to mislead the operator to estimate less load (80% the true value.)
total_Pload_diff = -0.2: When the attack is launched, the estimate of Pload is less than the estimate without attack by 20% of the true value. The attack is very successful in misleading operators to wrong estimate.